backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6
openai==1.57.2
langchain-text-splitters==0.3.2
google-auth-oauthlib==1.2.1
google-api-python-client==2.155.0
voyageai==0.3.2
twilio==9.3.7
resend==2.6.0
pymupdf==1.25.1
pytesseract==0.3.13
cryptography==43.0.3
pydantic-settings==2.7.0
pydantic==2.10.4
python-multipart==0.0.20
httpx==0.28.1
Pillow==11.0.0
